+++ blocks/rss-google-news-feature-block/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @wpmedia/rss-google-news-feature-block
 
+## 0.1.0-tk.0
+
+### Minor Changes
+
+- [`ecf8de2`](https://github.com/WPMedia/feed-components/commit/ecf8de222c38d0f06d0a2cf0f80caaaac75cdd18) [#90](https://github.com/WPMedia/feed-components/pull/90) Thanks [@tbrick855](https://github.com/tbrick855)! - Initial Commit
+
+### Patch Changes
+
+- Updated dependencies [[`50f362b`](https://github.com/WPMedia/feed-components/commit/50f362b3dd7979f904838755727360f10bfe07eb)]:
+  - @wpmedia/feeds-content-elements@0.3.0-tk.0
+  - @wpmedia/feeds-find-video-stream@0.1.0-tk.0
+
 ## 0.0.0
 
 ### Patch Changes